Move 'Shared items' related classes to own package
