+++ app/src/main/java/com/nextcloud/talk/dagger/modules/ViewModelModule.kt
@@ -23,7 +23,7 @@ package com.nextcloud.talk.dagger.modules
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
-import com.nextcloud.talk.viewmodels.SharedItemsViewModel
+import com.nextcloud.talk.shareditems.viewmodels.SharedItemsViewModel
 import dagger.Binds
 import dagger.MapKey
 import dagger.Module
